+++ src/main/ml-modules/root/config/searchResultsSortConfig.mjs
@@ -1,3 +1,6 @@
+import op from '/MarkLogic/optic';
+const lux = op.prefixer('https://lux.collections.yale.edu/ns/');
+
 const SORT_BINDINGS = {
   agentActiveDate: {
     indexType: 'field',
@@ -67,9 +70,29 @@ const SORT_BINDINGS = {
     indexReference: 'itemHeightDimensionValue',
     defaultOrder: 'asc',
   },
+  itemEncounterAgentName: {
+    predicate: lux('agentOfEncounter'),
+    indexReference: 'anySortName',
+    defaultOrder: 'asc',
+  },
+  itemEncounterPlaceName: {
+    predicate: lux('placeOfEncounter'),
+    indexReference: 'anySortName',
+    defaultOrder: 'asc',
+  },
   itemEncounteredDate: {
     indexType: 'field',
     indexReference: 'itemEncounteredStartDateLong',
+    defaultOrder: 'asc',
+  },
+  itemProductionAgentName: {
+    predicate: lux('agentOfProduction'),
+    indexReference: 'anySortName',
+    defaultOrder: 'asc',
+  },
+  itemProductionPlaceName: {
+    predicate: lux('placeOfProduction'),
+    indexReference: 'anySortName',
     defaultOrder: 'asc',
   },
   itemProductionDate: {